deletion protection

+++ Server/src/main/java/tfip/akimori/server/services/InstrumentService.java
@@ -184,15 +184,18 @@ public class InstrumentService {
         Instrument i = instruRepo.getInstrumentById(instrument_id);
         if (!storeRepo.isManagerOfStore(email, i.getStore_id())) {
             return false;
-        } else {
-            Boolean isDeleted = instruRepo.deleteInstrument(i);
-            if (isDeleted) {
-                String logMsg = "%s Deleted %s (S/N: %s) ".formatted(email, i.getInstrument_type(),
-                        i.getSerial_number());
-                logSvc.logInstrumentActivity(i.getStore_id(), "delete", email, i.getInstrument_id(), logMsg);
-                System.out.println("DELETED INSTRUMENT: " + i.getInstrument_id());
-            }
-            return isDeleted;
         }
+        if (i.getEmail() != null) {
+            return false;
+        }
+        Boolean isDeleted = instruRepo.deleteInstrument(i);
+        if (isDeleted) {
+            String logMsg = "%s Deleted %s (S/N: %s) ".formatted(email, i.getInstrument_type(),
+                    i.getSerial_number());
+            logSvc.logInstrumentActivity(i.getStore_id(), "delete", email, i.getInstrument_id(), logMsg);
+            System.out.println("DELETED INSTRUMENT: " + i.getInstrument_id());
+        }
+        return isDeleted;
+
     }
 }
